Moved score to a resources file.
